Implement A-Packages
Add minor bug fix which fixes.

Refactor code to package java classes into their own respective packages.
