--- src/main/java/Style.java
@@ -1,4 +1,0 @@
-public class Style {
-    static final String INDENTATION = "    ";
-    static final String HALF_INDENTATION = "  ";
-}
