+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package command;
+
+import storage.Storage;
+import task.TaskList;
+import ui.Ui;
+
 public class ExitCommand extends Command {
     public ExitCommand() {
         super(true);
