+++ src/main/java/exception/DukeException.java
@@ -1,5 +1,7 @@
+package exception;
+
 public class DukeException extends Exception {
-    DukeException(String message) {
+    public DukeException(String message) {
         super(message);
     }
