+++ src/main/java/task/Deadline.java
@@ -1,26 +1,27 @@
+package task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
+public class Deadline extends Task {
     protected LocalDate dateAndTime;
 
-    public Event(String deadline, String dateAndTime) {
+    public Deadline(String deadline, String dateAndTime) {
         super(deadline);
         this.dateAndTime = LocalDate.parse(dateAndTime);
     }
 
-    public Event(String deadline, String dateAndTime, boolean isDone) {
+    public Deadline(String deadline, String dateAndTime, boolean isDone) {
         super(deadline, isDone);
         this.dateAndTime = LocalDate.parse(dateAndTime);
     }
-
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: "
+        return "[D]" + super.toString() + " (by: "
                 + dateAndTime.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
     }
 
     public String toFileString() {
-        return "E , " + super.toFileString()  + " , " + dateAndTime;
+        return "D , " + super.toFileString()  + " , " + dateAndTime;
     }
 }
