+++ src/main/java/ui/Style.java
@@ -1,0 +1,6 @@
+package ui;
+
+public class Style {
+    public static final String INDENTATION = "    ";
+    public static final String HALF_INDENTATION = "  ";
+}
