+++ src/main/java/ui/Ui.java
@@ -1,4 +1,8 @@
-import java.util.ArrayList;
+package ui;
+
+import task.Task;
+import task.TaskList;
+
 import java.util.Scanner;
 public class Ui {
     boolean loadingError;
